좋아요기능추가

+++ back/src/db/models/Like.js
@@ -10,16 +10,21 @@ class like {
       const like = await likeModel.find({ adminId });
       return like;
     }
+    async findByUserId({ adminId, userId }) {
+        const like = await likeModel.findOne({ adminId, userId });
+        return like;
+    }
   
     async findByUserIdCount({ userId }) {
       const likes = await likeModel.find({ userId }).count();
       return likes;
     }
   
-    async deleteByid({ adminId }) {
-      const deleteLike = await likeModel.deleteOne({ adminId });
+    async deleteById({ adminId, userId }) {
+      const deleteLike = await likeModel.deleteOne({ adminId, userId });
       return deleteLike;
     }
   }
   const Like =  new like()
   export { Like };
+
